pyproject.toml
[project]
name = "hotel-management-systems"
version = "0.1.0"
description = ""
authors = [
    {name = "kashiftareen",email = "kashif.tareen1999@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.121.1,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry.scripts]
dev = "hotel_management_systems.main:start"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
